import restructuring, begin of test generalization

+++ src/spflow/torch/sampling/general/layers/leaves/parametric/cond_geometric.py
@@ -96,6 +96,6 @@ def sample(
         # data needs to be offset by +1 due to the different definitions between SciPy and PyTorch
         data[torch.meshgrid(sampling_ids, torch.tensor(node_scope.query), indexing="ij")] = (
             layer.dist(p=p, node_ids=[node_id]).sample((sampling_mask.sum(),)).to(p.device) + 1
-        )
+        ).type(torch.float64)
 
     return data
